fix(GiniHealthSDK): Refactor PaymentReviewViewController extensions in separate files + lang fixes
IPC-246
