+++ HealthSDK/GiniHealthSDK/Sources/GiniHealthSDK/Core/PaymentComponent/InstallAppBottomViewModel.swift
@@ -32,7 +32,7 @@ final class InstallAppBottomViewModel {
     let dimmingBackgroundColor: UIColor = GiniColor(lightModeColor: UIColor.black,
                                                     darkModeColor: UIColor.white).uiColor().withAlphaComponent(0.4)
 
-    var titleText: String = NSLocalizedStringPreferredFormat("ginihealth.paymentcomponent.installAppBottom.title",
+    var titleText: String = NSLocalizedStringPreferredFormat("ginihealth.paymentcomponent.installAppBottomSheet.title",
                                                              comment: "Install App Bottom sheet title")
     let titleLabelAccentColor: UIColor = GiniColor(lightModeColor: UIColor.GiniHealthColors.dark2,
                                                    darkModeColor: UIColor.GiniHealthColors.light2).uiColor()
@@ -55,11 +55,11 @@ final class InstallAppBottomViewModel {
                                                         darkModeColor: UIColor.GiniHealthColors.light3).uiColor()
     var moreInformationLabelText: String {
         isBankInstalled ? 
-        NSLocalizedStringPreferredFormat("ginihealth.paymentcomponent.installAppBottom.tip.description",
-                                         comment: "Text for tip information label").replacingOccurrences(of: bankToReplaceString, 
+        NSLocalizedStringPreferredFormat("ginihealth.paymentcomponent.installAppBottomSheet.tip.description",
+                                         comment: "Text for tip information label").replacingOccurrences(of: bankToReplaceString,
                                                                                                          with: selectedPaymentProvider?.name ?? "") :
-        NSLocalizedStringPreferredFormat("ginihealth.paymentcomponent.installAppBottom.notes.description",
-                                         comment: "Text for notes information label").replacingOccurrences(of: bankToReplaceString, 
+        NSLocalizedStringPreferredFormat("ginihealth.paymentcomponent.installAppBottomSheet.notes.description",
+                                         comment: "Text for notes information label").replacingOccurrences(of: bankToReplaceString,
                                                                                                            with: selectedPaymentProvider?.name ?? "")
     }
     
@@ -68,7 +68,7 @@ final class InstallAppBottomViewModel {
     let moreInformationIconName = "info.circle"
     
     // Pay invoice label
-    let continueLabelText: String = NSLocalizedStringPreferredFormat("ginihealth.paymentcomponent.installAppBottom.continue.button.text",
+    let continueLabelText: String = NSLocalizedStringPreferredFormat("ginihealth.paymentcomponent.installAppBottomSheet.continue.button.text",
                                                                      comment: "Title label used for the Continue button")
     
     var appStoreImageIconName = "appStoreIcon"
