+++ HealthSDK/GiniHealthSDK/Sources/GiniHealthSDK/Core/PaymentReviewViewController+UICollection.swift
@@ -1,0 +1,39 @@
+//
+//  PaymentReviewViewController+UICollection.swift
+//
+//  Copyright © 2024 Gini GmbH. All rights reserved.
+//
+
+import UIKit
+
+// MARK: - UICollectionViewDelegate, UICollectionViewDataSource
+
+extension PaymentReviewViewController: UICollectionViewDelegate, UICollectionViewDataSource, UICollectionViewDelegateFlowLayout {
+    public func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int { 1 }
+
+    public func numberOfSections(in collectionView: UICollectionView) -> Int {
+        model?.numberOfCells ?? 1
+    }
+
+    public func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
+        let cell = collectionView.dequeueReusableCell(withReuseIdentifier: "pageCellIdentifier", for: indexPath) as! PageCollectionViewCell
+        cell.pageImageView.frame = CGRect(x: 0, y: 0, width: collectionView.frame.width, height: collectionView.frame.height)
+        cell.pageImageView.contentInset = UIEdgeInsets(top: 0.0, left: 0.0, bottom: Constants.bottomPaddingPageImageView, right: 0.0)
+        let cellModel = model?.getCellViewModel(at: indexPath)
+        cell.pageImageView.display(image: cellModel?.preview ?? UIImage())
+        return cell
+    }
+    // MARK: - UICollectionViewDelegateFlowLayout
+
+    public func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, sizeForItemAt indexPath: IndexPath) -> CGSize {
+        let height = collectionView.frame.height
+        let width = collectionView.frame.width
+        return CGSize(width: width, height: height)
+    }
+
+    // MARK: - For Display the page number in page controll of collection view Cell
+
+    public func scrollViewDidScroll(_ scrollView: UIScrollView) {
+        pageControl.currentPage = Int(scrollView.contentOffset.x) / Int(scrollView.frame.width)
+    }
+}
