+++ HealthSDK/GiniHealthSDK/Sources/GiniHealthSDK/Resources/de.lproj/Localizable.strings
@@ -49,13 +49,13 @@
 "ginihealth.paymentcomponent.paymentinfo.questions.answer.clickable.text" = "Datenschutzerklärung[LINK]";
 "ginihealth.paymentcomponent.paymentinfo.gini.link" = "https://gini.net/";
 "ginihealth.paymentcomponent.paymentinfo.gini.privacypolicy.link" = "";
-"ginihealth.paymentcomponent.installAppBottom.title" = "Rechnungsdaten bereit zum Teilen mit der [BANK]";
-"ginihealth.paymentcomponent.installAppBottom.notes.description" = "Hinweis: Bitte aktualisieren oder installieren Sie zunächst die [BANK]-App aus dem App-Store.";
-"ginihealth.paymentcomponent.installAppBottom.tip.description" = "Tipp: Tippen Sie auf 'Weiter', um die Zahlung in der [BANK]-App abzuschließen.";
-"ginihealth.paymentcomponent.installAppBottom.continue.button.text" = "Weiter";
-"ginihealth.paymentcomponent.shareInvoiceBottom.title" = "Rechnungsdaten bereit zum Teilen mit der [BANK]";
-"ginihealth.paymentcomponent.shareInvoiceBottom.description" = "Im nächsten Schritt wählen Sie die [BANK]-App auf der Seite aus, um Ihre Zahlung in der [BANK]-App abzuschließen.";
-"ginihealth.paymentcomponent.shareInvoiceBottom.app" = "App";
-"ginihealth.paymentcomponent.shareInvoiceBottom.tip.description" = "Tipp: Scrollen Sie, bis Sie die DKB-App finden. Wenn Sie die DKB-App noch nicht installiert haben, laden Sie sie im Store herunter.";
-"ginihealth.paymentcomponent.shareInvoiceBottom.tip.underlined.part" = "laden Sie sie im Store herunter";
-"ginihealth.paymentcomponent.shareInvoiceBottom.continue.button.text" = "Weiter";
+"ginihealth.paymentcomponent.installAppBottomSheet.title" = "Rechnungsdaten bereit zum Teilen mit der [BANK]";
+"ginihealth.paymentcomponent.installAppBottomSheet.notes.description" = "Hinweis: Bitte aktualisieren oder installieren Sie zunächst die [BANK]-App aus dem App-Store.";
+"ginihealth.paymentcomponent.installAppBottomSheet.tip.description" = "Tipp: Tippen Sie auf 'Weiter', um die Zahlung in der [BANK]-App abzuschließen.";
+"ginihealth.paymentcomponent.installAppBottomSheet.continue.button.text" = "Weiter";
+"ginihealth.paymentcomponent.shareInvoiceBottomSheet.title" = "Rechnungsdaten bereit zum Teilen mit der [BANK]";
+"ginihealth.paymentcomponent.shareInvoiceBottomSheet.description" = "Im nächsten Schritt wählen Sie die [BANK]-App auf der Seite aus, um Ihre Zahlung in der [BANK]-App abzuschließen.";
+"ginihealth.paymentcomponent.shareInvoiceBottomSheet.app" = "App";
+"ginihealth.paymentcomponent.shareInvoiceBottomSheet.tip.description" = "Tipp: Scrollen Sie, bis Sie die DKB-App finden. Wenn Sie die DKB-App noch nicht installiert haben, laden Sie sie im Store herunter.";
+"ginihealth.paymentcomponent.shareInvoiceBottomSheet.tip.underlined.part" = "laden Sie sie im Store herunter";
+"ginihealth.paymentcomponent.shareInvoiceBottomSheet.continue.button.text" = "Weiter";
