+++ HealthSDK/GiniHealthSDK/Sources/GiniHealthSDK/Resources/en.lproj/Localizable.strings
@@ -49,13 +49,13 @@
 "ginihealth.paymentcomponent.paymentinfo.questions.answer.clickable.text" = "privacy policy[LINK]";
 "ginihealth.paymentcomponent.paymentinfo.gini.link" = "https://gini.net/en/";
 "ginihealth.paymentcomponent.paymentinfo.gini.privacypolicy.link" = "";
-"ginihealth.paymentcomponent.installAppBottom.title" = "Billing data ready to share with the [BANK] bank";
-"ginihealth.paymentcomponent.installAppBottom.notes.description" = "Note: You must first update or install the [BANK] app from the App store";
-"ginihealth.paymentcomponent.installAppBottom.tip.description" = "Tip: Tap 'Forward' to complete the payment. The [BANK] app will open automatically, for you to confirm the transaction. Thank you!";
-"ginihealth.paymentcomponent.installAppBottom.continue.button.text" = "Forward";
-"ginihealth.paymentcomponent.shareInvoiceBottom.title" = "Invoice data ready to share";
-"ginihealth.paymentcomponent.shareInvoiceBottom.description" = "In the next step, select the [BANK] app to open and complete your payment in the banking app.";
-"ginihealth.paymentcomponent.shareInvoiceBottom.app" = "App";
-"ginihealth.paymentcomponent.shareInvoiceBottom.tip.description" = "Tip: if you dont see the app scroll until you find the [BANK] app. If you haven't installed the [BANK] app yet, download it from the store.";
-"ginihealth.paymentcomponent.shareInvoiceBottom.tip.underlined.part" = "download it from the store";
-"ginihealth.paymentcomponent.shareInvoiceBottom.continue.button.text" = "Forward";
+"ginihealth.paymentcomponent.installAppBottomSheet.title" = "Billing data ready to share with the [BANK] bank";
+"ginihealth.paymentcomponent.installAppBottomSheet.notes.description" = "Note: You must first update or install the [BANK] app from the App store";
+"ginihealth.paymentcomponent.installAppBottomSheet.tip.description" = "Tip: Tap 'Forward' to complete the payment. The [BANK] app will open automatically, for you to confirm the transaction. Thank you!";
+"ginihealth.paymentcomponent.installAppBottomSheet.continue.button.text" = "Forward";
+"ginihealth.paymentcomponent.shareInvoiceBottomSheet.title" = "Invoice data ready to share";
+"ginihealth.paymentcomponent.shareInvoiceBottomSheet.description" = "In the next step, select the [BANK] app to open and complete your payment in the banking app.";
+"ginihealth.paymentcomponent.shareInvoiceBottomSheet.app" = "App";
+"ginihealth.paymentcomponent.shareInvoiceBottomSheet.tip.description" = "Tip: if you dont see the app scroll until you find the [BANK] app. If you haven't installed the [BANK] app yet, download it from the store.";
+"ginihealth.paymentcomponent.shareInvoiceBottomSheet.tip.underlined.part" = "download it from the store";
+"ginihealth.paymentcomponent.shareInvoiceBottomSheet.continue.button.text" = "Forward";
